Fixed more code
Testing out a new function testCollision. Turns out the circles were not interacting with each other. I fixed the intersection code, now they have intersection detection, still is very jank, it changes the colors of the circle on collision, but I also implemented "testCollision" and got the circles to actually INTERACT with each other, which is HUGE.

TODO
- Figure out why resolveCollision isn't working as intended || not responsive
- Resolve the collision so they the circles collide and react according to huma intuition.
- Find out why the collision detection